+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
-## v2.0.0-beta.6
-- **[fix]** - module augmentions export
+## v2.0.0-beta.7
+- **[fix]** - module augmentions
+- **[fix]** - remove unused navigation props